# data/swagger.json
{
   "swagger": "2.0",
   "info": {
      "description": "API for Fixthis",
      "version": "0.1.0",
      "title": "Fixthis",
      "termsOfService": ""
   },
   "basePath": "/api",
...
   "schemes": [
      "http"
   ],
   "paths": {
...
      "/auth/social": {
         "post": {
            "tags": [
               "Authentication"
            ],
            "summary": "Authenticate with social media",
            "description": "Api to authenticate with social media",
            "operationId": "auth.authSocial",
            "consumes": [
               "application/json"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "body",
                  "name": "authSocialRequestObj",
                  "required": true,
                  "description": "Request Object",
                  "schema": {
                     "$ref": "#/definitions/LoginSocialRequest"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/auth/forgotpassword": {
         "post": {
            "tags": [
               "Authentication"
            ],
            "summary": "Forgot Password",
            "description": "Api for forgot password.",
            "operationId": "auth.forgotpassword",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "email",
                  "required": true,
                  "description": "User email",
                  "type": "string"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/auth/resetpassword/{token}": {
         "post": {
            "tags": [
               "Authentication"
            ],
            "summary": "Verify user account",
            "description": "Api to reset user password",
            "operationId": "api.resetpassword",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "path",
                  "name": "token",
                  "type": "string",
                  "required": true,
                  "description": "Reset token"
               },
               {
                  "in": "formData",
                  "name": "password",
                  "type": "string",
                  "required": true,
                  "description": "Password to reset"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
   },
   "definitions": {
...
      "LoginSocialRequest": {
         "type": "object",
         "required": [
            "social_site",
            "site_auth_token"
         ],
         "properties": {
            "social_site": {
               "type": "string",
               "default": "fb/google"
            },
            "site_auth_token": {
               "type": "string",
               "format": "fb_auth_token"
            },
            "device_token": {
               "type": "string",
               "format": "string"
            },
            "device_type": {
               "type": "string",
               "format": "string"
            }
         }
      },
...
   },
   "securityDefinitions": {
      "bearer": {
         "in": "header",
         "name": "Authorization",
         "type": "apiKey"
      }
   }
}